rpi_gateway/requirements.txt
# M.A.S.H. IoT - Python Dependencies
# Raspberry Pi Gateway Requirements

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Serial Communication
pyserial==3.5

# Database
# SQLite is built into Python, no separate package needed

# Firebase Integration
firebase-admin==6.3.0
python-firebase==1.2

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
joblib==1.3.2

# System & Utilities
python-dotenv==1.0.0
PyYAML==6.0.1

# Networking
requests==2.31.0

# MQTT (HiveMQ)
paho-mqtt==1.6.1

# Logging & Monitoring
colorlog==6.8.0

# Development Tools (optional)
pytest==7.4.3
pytest-flask==1.3.0
